Move image filter to its own component
